feat: wip refactor
