+++ packages/core/src/index.ts
@@ -1,11 +1,5 @@
-import Buffer from 'buffer'
-import CID from 'cids'
-
 export * from './thread'
 export * from './service'
 export * from './external'
-
-export interface Block {
-  data: Buffer
-  cid: CID
-}
+export * from './utils'
+export * from './store'
