+++ packages/core/src/store/query.ts
