+++ packages/core/src/utils/index.ts
@@ -1,4 +1,11 @@
-import { ThreadID, LogID } from '@textile/threads-core'
+import Buffer from 'buffer'
+import CID from 'cids'
+import { ThreadID, LogID } from '../thread'
+
+export interface Block {
+  data: Buffer
+  cid: CID
+}
 
 export interface LogsThreads {
   // logs returns a list of log IDs for a thread.
