+++ packages/service-client/README.md
@@ -1,0 +1,57 @@
+# Threads Network Client _(threads-network-client)_
+
+[![Made by Textile](https://img.shields.io/badge/made%20by-Textile-informational.svg?style=popout-square)](https://textile.io)
+[![Chat on Slack](https://img.shields.io/badge/slack-slack.textile.io-informational.svg?style=popout-square)](https://slack.textile.io)
+[![npm (scoped)](https://img.shields.io/npm/v/@textile/threads-network-client.svg?style=popout-square)](https://www.npmjs.com/package/@textile/threads-network-client)
+[![docs](https://img.shields.io/badge/docs-master-success.svg?style=popout-square)](https://textileio.github.io/js-threads)
+[![standard-readme compliant](https://img.shields.io/badge/standard--readme-OK-green.svg?style=flat-square)](https://github.com/RichardLitt/standard-readme)
+
+> Client for interacting with Textile's (remote) Threads gRPC Service
+
+Join us on our [public Slack channel](https://slack.textile.io/) for news, discussions, and status updates. [Check out our blog](https://blog.textile.io) for the latest posts and announcements.
+
+## Table of Contents
+
+- [Background](#background)
+- [Install](#install)
+- [Usage](#usage)
+- [API](#api)
+- [Maintainers](#maintainers)
+- [Contributing](#contributing)
+- [License](#license)
+
+## Background
+
+TODO
+
+## Install
+
+```bash
+// TODO
+```
+
+## Usage
+
+```typescript
+// TODO
+```
+
+## API
+
+See [https://textileio.github.io/js-threads](https://textileio.github.io/js-threads)
+
+## Maintainers
+
+[Carson Farmer](https://github.com/carsonfarmer)
+
+## Contributing
+
+See [the contributing file](../../CONTRIBUTING.md)!
+
+PRs accepted.
+
+Small note: If editing the README, please conform to the [standard-readme](https://github.com/RichardLitt/standard-readme) specification.
+
+## License
+
+[MIT](../../LICENSE) (c) 2019 Textile
