+++ packages/service/src/index.ts
@@ -1,3 +1,2 @@
 export * from './logstore'
-export * from './network'
 export * from './service'
