+++ packages/service/src/logstore/metadatabook.ts
@@ -1,9 +1,8 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 import { Datastore, Key } from 'interface-datastore'
 import { NamespaceDatastore } from 'datastore-core'
-import { ThreadID } from '@textile/threads-core'
+import { ThreadID, Closer } from '@textile/threads-core'
 import { encode, decode } from 'cbor-sync'
-import { Closer } from './interface'
 
 // /thread/heads/<base32 thread id no padding>/<base32 log id no padding>
 const baseKey = new Key('/thread/meta')
