+++ packages/store/src/codecs/index.ts
@@ -1,4 +1,3 @@
 import * as JSONPatcher from './jsonpatcher'
 
-export * from './interface'
 export { JSONPatcher }
