+++ packages/store/src/dispatcher/index.spec.ts
@@ -1,7 +1,7 @@
 import { expect } from 'chai'
 import { pack } from 'lexicographic-integer'
-import { Dispatcher } from './dispatcher'
-import { Event } from '..'
+import { Event } from '@textile/threads-core'
+import { Dispatcher } from '.'
 
 describe('Dispatcher', () => {
   it('should not require any arguments to initialise', async () => {
